Fix indel phasing: HP was never assigned to indels
The problem was that variants are identified by position and normalization
changes the position. That is, when the PhasedVcfWriter looked up the phase
of the variant, it could never find it.

With this change, indel normalization is repeated in the PhasedVcfWriter
before looking up the variant. (The output variant does not change.)