Read full chromosome sequence before re-aligning reads to it
The reference sequence was passed as a pyfaidx.FastaRecord object to the
re-alignment function. Slicing a FastaRecord is quite fast, but slicing a
regular str is even faster.

This change increases memory usage of WhatsHap by the size of the largest
processed chromosome, but reduces runtime by a few percent.