20% speedup by accessing AlignedSegment.cigartuples only once
AlignedSegment.cigartuples is a property that gets re-computed every
time it is accessed.  We were accessing it twice, creating quite a long
list of tuples every time.  Avoiding this requires adding an ugly
cigartuples parameter to a few methods, but it is worth it.

Reading in the BAM file and detecting alleles in it now takes 160% of the
time it takes to just run 'samtools view' on the BAM file.