Support writing compressed output
The output file (-o/--output) of the phase command is now opened with
xopen, which supports .gz, .bz2 and .xz compression, auto-detected
from the file name.

More useful would be to compress the output with bgzip such that it can be
indexed with tabix, or to directly support .bcf output.