Switch to pysam.VariantFile for reading and writing VCFs
- Also add a VcfAugmenter class that is a base class for the two VCF writers.

- We no longer need the xopen and PyVCF dependencies.

- We can now have .bcf output (untested)

- .vcf.gz output is now compressed with bgzip, so tabix can be used.

Problems:

VariantFile sometimes segfaults when the input file that we augment with
phasing information does not contain all the necessary INFO/FORMAT/contig
header fields. They are optional in the file itself according to VCF spec,
but htslib (on which VariantFile is based) needs to know the correct types
for the tags that it writes.