Handle input VCFs with incomplete headers
pysam.VariantFile needs all fields that are written to an output file
(contigs, FORMATs, INFOs) to be defined in the header and segfaults
otherwise. To prevent this, we now go through the entire input VCF once,
collecting all the contigs, FORMATs, INFOs that are used, and add them to
the header. This fails if we non-standard tags are used, but we can at least
detect this and not crash.