Reduce memory usage when reading VCF with many samples
Instead of using our own function, which reads in a full chromosome at a
time and stores all records in a list, use itertools.groupby instead, which
yields iterators. That is, the original record list is no longer kept in
memory.

This is still very slow and may still have a high memory usage, but it’s at
least better than before.

See issue #86