Avoid importing pkg_resources, it is slow
Instead, use the setuptools_scm feature of generating a _version.py file and
import that instead

I’ve done the same change in cutadapt and xopen in order to reduce import
time (and therefore startup time for command-line tools). WhatsHap startup
time is still large, so it doesn’t help that much.