Introduce CommandLineError exception
Instead of calling sys.exit directly in each of the run_* functions, we now
throw a CommandLineError. This shortens the code a little bit, but allows us
to limit the scope of some unit tests. Instead of checking whether a
SystemExit occurred, we can now check whether a CommandLineError occurred.

If a CommandLineError occurs, an error message is printed and the program
exits with an exit code of 1.