Handle unknown alleles in phased genotypes
If the GT in the input VCF looks like "1|." or similar, this is encoded as a
VariantCallPhase with phase=(1, None), and when converting all alleles of
the second haplotype in the phase set to a string, this ends up being
encoded as the literal string "None" in the string that encodes the phasing.
This crashes at the latest when the hamming distance is computed, which
needs both arguments to be of the same length.

Closes #342