Document what a block N50 of 0 means
We already use "nan" for the case that no contig lengths are availeble. In
order to be able to distinguish that from the case that the block lengths do
not sum up to half the contig lengths, it makes sense to use a different
value, so let’s just document the current behavior, which is to write a 0.

Closes #219